deepfake-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
timm>=1.0.0
transformers>=4.40.0
grad-cam>=1.5.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
requests>=2.31.0
groq>=0.5.0
